moved the query builder logic to a shared place
